+++ R/engines.R
@@ -83,6 +83,8 @@ use_epoxy_knitr_engines <- function(
 		use_epoxy_glue_engine()
 	}
 
+	knitr_register_detect_inline()
+
 	invisible(old)
 }
 
@@ -122,7 +124,7 @@ prefer_dotted_data_option <- function(options) {
 
 eval_epoxy_engine <- function(fn, code, options) {
 	defaults <- formals(fn)
-	exclude <- c("...", ".data", ".style", ".transformer")
+	exclude <- c("...", ".data", ".style", ".transformer", ".envir")
 	defaults <- defaults[setdiff(names(defaults), exclude)]
 	defaults <- lapply(defaults, rlang::eval_bare, env = environment(fn))
 	defaults$.envir <- knitr::knit_global()